Pipe/requirements.txt
dotmap==1.3.30
einops==0.7.0
einops_exts==0.0.4
ftfy==6.1.3
fvcore==0.1.5.post20221221
ipdb==0.13.13
matplotlib==3.3.4
mmcv_full==1.7.1
numpy==1.23.5
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.5.0
Pillow==10.2.0
PyYAML==6.0
PyYAML==6.0.1
RandAugment==0.1
RandAugment==1.0.2
regex==2021.7.6
scikit_learn==0.22.1
scipy==1.6.1
setuptools==60.2.0
timm==0.6.12
torch==2.0.1+cu117
torchvision==0.15.2+cu117
tqdm==4.65.2
transformers==4.28.1
